fin back messagerie

+++ app/Http/Controllers/API/AuthController.php
@@ -21,132 +21,132 @@ use Illuminate\Support\Facades\Log; // Importez la classe Log
 class AuthController extends Controller
 {
     public function register(Request $request)
-    {
-        $validator = Validator::make($request->all(), [
-            'nom' => 'required|string',
-            'prenom' => 'required|string',
-            'ville' => 'required|string',
-            'date_naissance' => 'required|string',
-            'numero_tel' => 'required|string',
-            'type_compte' => 'required|string',
-            'photo' => 'nullable|string',
-            'email' => 'nullable|email|unique:users,email',
-            'password' => 'nullable|string|min:6',
-        ]);
-    
-        // Vérifier si la validation a échoué
-        if ($validator->fails()) {
-            return response()->json(['errors' => $validator->errors()], 422);
+        {
+            $validator = Validator::make($request->all(), [
+                'nom' => 'required|string',
+                'prenom' => 'required|string',
+                'ville' => 'required|string',
+                'date_naissance' => 'required|string',
+                'numero_tel' => 'required|string',
+                'type_compte' => 'required|string',
+                'photo' => 'nullable|string',
+                'email' => 'nullable|email|unique:users,email',
+                'password' => 'nullable|string|min:6',
+            ]);
+        
+            // Vérifier si la validation a échoué
+            if ($validator->fails()) {
+                return response()->json(['errors' => $validator->errors()], 422);
+            }
+        
+            $randomNumber = random_int(10000000, 99999999);
+        
+            // Création de l'utilisateur dans la base de données
+            $user = User::create([
+                'nom' => $request->nom,
+                'prenom' => $request->prenom,
+                'ville' => $request->ville,
+                'date_naissance' => $request->date_naissance,
+                'numero_tel' => $request->numero_tel,
+                'type_compte' => $request->type_compte,
+                'otp' => $randomNumber,
+                'photo' => $request->photo,
+                'email' => $request->email,
+                'password' => $request->password ? Hash::make($request->password) : null,
+            ]);
+        
+            // Sauvegarde du type_compte et du token dans un fichier JSON
+            $userData = [
+                'type_compte' => $user->type_compte,
+                'token' => $user->token,
+            ];
+        
+            // Chemin vers le fichier JSON
+            $jsonFilePath = storage_path('app/user_data.json');
+        
+            // Vérifier si le fichier JSON existe
+            if (file_exists($jsonFilePath)) {
+                // Lire le contenu actuel du fichier JSON
+                $currentData = json_decode(file_get_contents($jsonFilePath), true);
+                // Ajouter les nouvelles données utilisateur
+                $currentData[] = $userData;
+                // Écrire les données mises à jour dans le fichier JSON
+                file_put_contents($jsonFilePath, json_encode($currentData));
+            } else {
+                // Créer un nouveau fichier JSON avec les données utilisateur
+                file_put_contents($jsonFilePath, json_encode([$userData]));
+            }
+        
+            // Créer le jeton d'authentification
+            $token = $user->createToken('auth_token')->plainTextToken;
+        
+            // Ajouter un log pour afficher le jeton d'authentification
+            Log::info('Token created: ' . $token);
+        
+            return response()->json([
+                'message' => 'Utilisateur enregistré avec succès.',
+                'role' => $user->type_compte,
+                'token' => $token,
+            ], 201);
         }
-    
-        $randomNumber = random_int(10000000, 99999999);
-    
-        // Création de l'utilisateur dans la base de données
-        $user = User::create([
-            'nom' => $request->nom,
-            'prenom' => $request->prenom,
-            'ville' => $request->ville,
-            'date_naissance' => $request->date_naissance,
-            'numero_tel' => $request->numero_tel,
-            'type_compte' => $request->type_compte,
-            'otp' => $randomNumber,
-            'photo' => $request->photo,
-            'email' => $request->email,
-            'password' => $request->password ? Hash::make($request->password) : null,
-        ]);
-    
-        // Sauvegarde du type_compte et du token dans un fichier JSON
-        $userData = [
-            'type_compte' => $user->type_compte,
-            'token' => $user->token,
-        ];
-    
-        // Chemin vers le fichier JSON
-        $jsonFilePath = storage_path('app/user_data.json');
-    
-        // Vérifier si le fichier JSON existe
-        if (file_exists($jsonFilePath)) {
-            // Lire le contenu actuel du fichier JSON
-            $currentData = json_decode(file_get_contents($jsonFilePath), true);
-            // Ajouter les nouvelles données utilisateur
-            $currentData[] = $userData;
-            // Écrire les données mises à jour dans le fichier JSON
-            file_put_contents($jsonFilePath, json_encode($currentData));
-        } else {
-            // Créer un nouveau fichier JSON avec les données utilisateur
-            file_put_contents($jsonFilePath, json_encode([$userData]));
-        }
-    
-        // Créer le jeton d'authentification
-        $token = $user->createToken('auth_token')->plainTextToken;
-    
-        // Ajouter un log pour afficher le jeton d'authentification
-        Log::info('Token created: ' . $token);
-    
-        return response()->json([
-            'message' => 'Utilisateur enregistré avec succès.',
-            'role' => $user->type_compte,
-            'token' => $token,
-        ], 201);
-    }
 
     public function enregistrerCamion(Request $request)
-{
-        // Verify user authentication
-        if (!Auth::check()) {
-            return response()->json([
-                'message' => 'Accès non autorisé. Veuillez vous connecter d\'abord.',
-            ], 401);
+        {
+                // Verify user authentication
+                if (!Auth::check()) {
+                    return response()->json([
+                        'message' => 'Accès non autorisé. Veuillez vous connecter d\'abord.',
+                    ], 401);
+                }
+
+                // Récupérer le numéro de téléphone de l'utilisateur connecté
+                $user = Auth::user();
+                $numeroTel = $user->numero_tel;
+
+                // Valider les données entrantes
+                $validatedData = $request->validate([
+                    'matricule' => 'required|string|unique:vehicules,matricule',
+                    'photo_camion' => 'required|file|mimetypes:image/jpeg,image/png,image/jpg,image/gif,application/pdf|max:2048',
+                    'carte_grise' => 'required|file|mimetypes:image/jpeg,image/png,image/jpg,image/gif,application/pdf|max:2048',
+                    'visite_technique' => 'required|file|mimetypes:image/jpeg,image/png,image/jpg,image/gif,application/pdf|max:2048',
+                    'assurance' => 'required|file|mimetypes:image/jpeg,image/png,image/jpg,image/gif,application/pdf|max:2048',
+
+            ]);
+
+
+            try {
+                // Enregistrement des données dans la base de données
+                $enregistrement = new Enregistrement();
+                $enregistrement->matricule = $validatedData['matricule'];
+                $enregistrement->numero_tel = $numeroTel;
+                $enregistrement->statut = 'En attent';
+
+                // Stocker les fichiers téléchargés localement
+                $photoCamion = $request->file('photo_camion');
+                $carteGrise = $request->file('carte_grise');
+                $visiteTechnique = $request->file('visite_technique');
+                $assurance = $request->file('assurance');
+
+                // Utiliser le nom original des fichiers lors du stockage
+                $photoPath = $photoCamion->storeAs('public/images', $photoCamion->getClientOriginalName());
+                $carteGrisePath = $carteGrise->storeAs('public/images', $carteGrise->getClientOriginalName());
+                $visiteTechniquePath = $visiteTechnique->storeAs('public/images', $visiteTechnique->getClientOriginalName());
+                $assurancePath = $assurance->storeAs('public/images', $assurance->getClientOriginalName());
+
+                // Mettre à jour les chemins complets dans l'objet Enregistrement
+                $enregistrement->photo_camion = str_replace('public/', '', $photoPath);
+                $enregistrement->carte_grise = str_replace('public/', '', $carteGrisePath);
+                $enregistrement->visite_technique = str_replace('public/', '', $visiteTechniquePath);
+                $enregistrement->assurance = str_replace('public/', '', $assurancePath);
+
+                $enregistrement->save();
+
+                return response()->json(['message' => 'Enregistrement réussi', 'data' => $enregistrement], 201);
+            }  catch (\Exception $e) {
+                Log::error('Erreur lors de l\'enregistrement du camion: ' . $e->getMessage());
+                return response()->json(['message' => 'Erreur lors de l\'enregistrement du camion'], 500);
+            }
         }
-
-        // Récupérer le numéro de téléphone de l'utilisateur connecté
-        $user = Auth::user();
-        $numeroTel = $user->numero_tel;
-
-        // Valider les données entrantes
-        $validatedData = $request->validate([
-            'matricule' => 'required|string|unique:vehicules,matricule',
-            'photo_camion' => 'required|file|mimetypes:image/jpeg,image/png,image/jpg,image/gif,application/pdf|max:2048',
-            'carte_grise' => 'required|file|mimetypes:image/jpeg,image/png,image/jpg,image/gif,application/pdf|max:2048',
-            'visite_technique' => 'required|file|mimetypes:image/jpeg,image/png,image/jpg,image/gif,application/pdf|max:2048',
-            'assurance' => 'required|file|mimetypes:image/jpeg,image/png,image/jpg,image/gif,application/pdf|max:2048',
-
-    ]);
-
-
-    try {
-        // Enregistrement des données dans la base de données
-        $enregistrement = new Enregistrement();
-        $enregistrement->matricule = $validatedData['matricule'];
-        $enregistrement->numero_tel = $numeroTel;
-        $enregistrement->statut = 'En attent';
-
-        // Stocker les fichiers téléchargés localement
-        $photoCamion = $request->file('photo_camion');
-        $carteGrise = $request->file('carte_grise');
-        $visiteTechnique = $request->file('visite_technique');
-        $assurance = $request->file('assurance');
-
-        // Utiliser le nom original des fichiers lors du stockage
-        $photoPath = $photoCamion->storeAs('public/images', $photoCamion->getClientOriginalName());
-        $carteGrisePath = $carteGrise->storeAs('public/images', $carteGrise->getClientOriginalName());
-        $visiteTechniquePath = $visiteTechnique->storeAs('public/images', $visiteTechnique->getClientOriginalName());
-        $assurancePath = $assurance->storeAs('public/images', $assurance->getClientOriginalName());
-
-        // Mettre à jour les chemins complets dans l'objet Enregistrement
-        $enregistrement->photo_camion = str_replace('public/', '', $photoPath);
-        $enregistrement->carte_grise = str_replace('public/', '', $carteGrisePath);
-        $enregistrement->visite_technique = str_replace('public/', '', $visiteTechniquePath);
-        $enregistrement->assurance = str_replace('public/', '', $assurancePath);
-
-        $enregistrement->save();
-
-        return response()->json(['message' => 'Enregistrement réussi', 'data' => $enregistrement], 201);
-    }  catch (\Exception $e) {
-        Log::error('Erreur lors de l\'enregistrement du camion: ' . $e->getMessage());
-        return response()->json(['message' => 'Erreur lors de l\'enregistrement du camion'], 500);
-    }
-}
 
 
 public function mettreAJourCamion(Request $request, $matricule)
@@ -181,7 +181,7 @@ public function mettreAJourCamion(Request $request, $matricule)
         // Gérer les mises à jour des fichiers
         if ($request->hasFile('photo_camion')) {
             // Supprimer l'ancienne image si elle existe
-            if ($enregistrement->photo_camion) {
+            if ($enregistrement->photo_camion && Storage::exists('public/' . $enregistrement->photo_camion)) {
                 Storage::delete('public/' . $enregistrement->photo_camion);
             }
             // Enregistrer la nouvelle image
@@ -191,7 +191,7 @@ public function mettreAJourCamion(Request $request, $matricule)
         }
 
         if ($request->hasFile('carte_grise')) {
-            if ($enregistrement->carte_grise) {
+            if ($enregistrement->carte_grise && Storage::exists('public/' . $enregistrement->carte_grise)) {
                 Storage::delete('public/' . $enregistrement->carte_grise);
             }
             $carteGrise = $request->file('carte_grise');
@@ -200,7 +200,7 @@ public function mettreAJourCamion(Request $request, $matricule)
         }
 
         if ($request->hasFile('visite_technique')) {
-            if ($enregistrement->visite_technique) {
+            if ($enregistrement->visite_technique && Storage::exists('public/' . $enregistrement->visite_technique)) {
                 Storage::delete('public/' . $enregistrement->visite_technique);
             }
             $visiteTechnique = $request->file('visite_technique');
@@ -209,7 +209,7 @@ public function mettreAJourCamion(Request $request, $matricule)
         }
 
         if ($request->hasFile('assurance')) {
-            if ($enregistrement->assurance) {
+            if ($enregistrement->assurance && Storage::exists('public/' . $enregistrement->assurance)) {
                 Storage::delete('public/' . $enregistrement->assurance);
             }
             $assurance = $request->file('assurance');
@@ -219,7 +219,7 @@ public function mettreAJourCamion(Request $request, $matricule)
 
         $enregistrement->save();
 
-        return response()->json(['message' => 'Camion mis à jour avec succès', 'data' => $enregistrement], 201);
+        return response()->json(['message' => 'Camion mis à jour avec succès', 'data' => $enregistrement], 200);
     } catch (\Exception $e) {
         Log::error('Erreur lors de la mise à jour du camion: ' . $e->getMessage());
         return response()->json(['message' => 'Erreur lors de la mise à jour du camion'], 500);
@@ -418,19 +418,26 @@ public function getCamionDetails(Request $request, $matricule)
                 // Récupérer les messages envoyés par cet utilisateur
                 $sentMessages = Envoyer::where('numero_tel', $user->numero_tel)
                     ->orderBy('created_at', 'asc')
-                    ->get();
-
-                // Récupérer les messages reçus par cet utilisateur
+                    ->get()
+                    ->map(function ($message) {
+                        $message->type = 'sent';
+                        return $message;
+                    });
+    
+                // Récupérer les messages reçus par cet utilisateur (qui sont de l'admin)
                 $receivedMessages = Recevoir::where('numero_tel', $user->numero_tel)
                     ->orderBy('created_at', 'asc')
-                    ->get();
-
-              // Fusionner les deux collections et les convertir en tableau
-            $messages = $sentMessages->merge($receivedMessages)->sortBy('created_at')->values()->all();
-
+                    ->get()
+                    ->map(function ($message) {
+                        $message->type = 'admin';
+                        return $message;
+                    });
+    
+                // Fusionner les deux collections et les convertir en tableau
+                $messages = $sentMessages->merge($receivedMessages)->sortBy('created_at')->values()->all();
     
                 // Retourner les messages envoyés par l'utilisateur sous forme de réponse JSON
-                return response()->json(['messages' => [$messages]], 201);
+                return response()->json(['messages' => $messages], 201);
             } else {
                 // Retourner une réponse indiquant que l'utilisateur n'est pas trouvé
                 return response()->json(['error' => 'Utilisateur non trouvé'], 404);
@@ -440,6 +447,7 @@ public function getCamionDetails(Request $request, $matricule)
             return response()->json(['error' => 'Échec de la récupération des messages'], 500);
         }
     }
+    
     
 
   /*   public function receiveAdmin(Request $request)
